Refactoring, new Assertions-specific test suite
Since Assertions are now in their own TU, they now have their own
test module.
Assertion tests are removed from TestSuite tests.

Removed now-unnecessary TestCPPUtil.h #include.
Refactored assertion calls.
Formatting.
