+++ test/include/Assertions/AssertionsTests.h
@@ -1,0 +1,15 @@
+#ifndef TESTCPP_ASSERTIONS_TESTS_
+#define TESTCPP_ASSERTIONS_TESTS_
+
+namespace TestCPP {
+    namespace Testing {
+        namespace AssertionsTests {
+            void TestAssertTrue ();
+            void TestAssertFalse ();
+            void TestAssertNull ();
+            void TestAssertNotNull ();
+        }
+    }
+}
+
+#endif
