+++ test/include/TestSuite/TestSuiteSuite.h
@@ -12,22 +12,6 @@ namespace TestCPP {
                 make_tuple(
                     "Suite construction Test",
                     function<void()>(TestSuiteTests::TestConstructSuite)
-                ),
-                make_tuple(
-                    "assertNull Test",
-                    function<void()>(TestSuiteTests::TestAssertNull)
-                ),
-                make_tuple(
-                    "assertNotNull Test",
-                    function<void()>(TestSuiteTests::TestAssertNotNull)
-                ),
-                make_tuple(
-                    "assertTrue Test",
-                    function<void()>(TestSuiteTests::TestAssertTrue)
-                ),
-                make_tuple(
-                    "assertFalse Test",
-                    function<void()>(TestSuiteTests::TestAssertFalse)
                 )
             );
         }
